refactor : 메인페이지 컴포턴트 분류
