+++ src/pages/MainPage.js
@@ -1,22 +1,13 @@
 import { useNavigate } from "react-router-dom";
-import styles from "./MainPage.module.css";
-
+import styles from "../components/MainPage.module.css";
 import styled from "styled-components";
 
-import profileSample from "../img/profileSample.png";
 import HeaderComponent from "../components/MainPage/HeaderComponent";
+import { ProfileComponent } from "../components/MainPage/ProfileComponent";
 import { Link } from "react-scroll";
 import { useEffect, useState } from "react";
 
-const ProfilePicture = ({ src }) => {
-  return (
-    <>
-      <img src={src} alt="profImg" />
-    </>
-  );
-};
-
-function MainPage() {
+export default function MainPage() {
   const CheckLogin = () => {
     if (!localStorage.getItem("loginToken")) {
       alert("문제가 발생했습니다. 로그인 페이지로 이동합니다");
@@ -24,11 +15,6 @@ function MainPage() {
       navigate("/");
     }
   };
-  useEffect(() => {
-    console.log("CHECK LOGIN");
-    CheckLogin();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
 
   const navigate = useNavigate();
   const [post, setPost] = useState({
@@ -57,6 +43,7 @@ function MainPage() {
         throw new Error(`HTTP 에러 Status: ${response.status}`);
       }
       const responseData = await response.json();
+
       setPost({
         firstName: responseData.firstName,
         lastName: responseData.lastName,
@@ -71,7 +58,9 @@ function MainPage() {
   };
 
   useEffect(() => {
+    CheckLogin();
     getProfile();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   return (
@@ -91,41 +80,13 @@ function MainPage() {
           <p>나의 이력을 관리하고 원하는 팀을 구성할 수 있어요</p>
         </div>
         <input placeholder="원하는 공모전, 대회 등을 입력해보세요" />
-        <div className={styles.profile}>
-          <>
-            <>
-              {(!post?.profileImageUrl && post?.profileImageUrl === null) ||
-              post?.profileImageUrl === undefined ? (
-                <>
-                  {console.log("no profile", post?.profileImageUrl)}
-                  <ProfilePicture src={profileSample} />
-                </>
-              ) : (
-                <>
-                  <ProfilePicture src={post?.profileImageUrl} />
-                </>
-              )}
-            </>
-          </>
-          <TextWrapper>
-            <p className={styles.name}>
-              {post.firstName} {post.lastName}
-            </p>
-            {/* <Wrapper> */}
-            <p className={styles.academicInfo1}>
-              한동대학교 {post.department} {post.semester}학기
-            </p>
-            <p className={styles.academicInfo}>{post.email}</p>
-            {/* </Wrapper> */}
-          </TextWrapper>
-        </div>
         <Link to="mainImg" spy={true} smooth={true}>
           <button className={styles.move} style={{ cursor: "pointer" }}>
             <img src="img/move.png" alt="img" />
           </button>
         </Link>
       </div>
-
+      <ProfileComponent post={post} />
       <div className={styles.grid}>
         <img src="img/mainImg2.png" alt="img" />
         <img src="img/mainImg.png" alt="img" id="mainImg" />
@@ -136,19 +97,6 @@ function MainPage() {
     </div>
   );
 }
-
-export default MainPage;
-
-const TextWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-  > p {
-    margin-bottom: 0px;
-  }
-`;
 
 // const Wrapper = styled.div`
 //   margin-top: 10px;
